requirements (1).txt
streamlit
sentence-transformers
faiss-cpu
feedparser
requests
